packages/agents/pyproject.toml
[project]
name = "jurisflow-agents"
version = "0.1.0"
description = "Google ADK workflow definitions for Jurisflow."
requires-python = ">=3.11"
dependencies = [
  "google-adk>=1.27,<2.0",
  "httpx>=0.28,<1.0",
  "jurisflow-db",
  "jurisflow-retrieval",
  "jurisflow-shared",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
